feat: 회원 탈퇴 기능

+++ src/main/java/com/kulift/lift/user/controller/UserController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -13,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.kulift.lift.auth.security.CustomUserDetails;
+import com.kulift.lift.global.exception.CustomException;
+import com.kulift.lift.global.exception.ErrorCode;
 import com.kulift.lift.user.dto.PasswordUpdateRequest;
 import com.kulift.lift.user.dto.UserResponse;
 import com.kulift.lift.user.service.UserService;
@@ -51,8 +54,18 @@ public class UserController {
 	}
 
 	@DeleteMapping("/{id}")
-	public ResponseEntity<String> deleteUser(@PathVariable Long id) {
+	public ResponseEntity<String> deleteUser(@AuthenticationPrincipal CustomUserDetails userDetails,
+		@PathVariable Long id) {
+		if (!userDetails.getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))) {
+			throw new CustomException(ErrorCode.FORBIDDEN);
+		}
 		userService.delete(id);
 		return ResponseEntity.ok("삭제되었습니다.");
+	}
+
+	@DeleteMapping
+	public ResponseEntity<String> withdrawUser(@AuthenticationPrincipal CustomUserDetails userDetails) {
+		userService.delete(userDetails.getId());
+		return ResponseEntity.ok("회원 탈퇴 완료.");
 	}
 }
